pre push version 1

+++ src/main/java/kz/syllabus/entity/User.java
@@ -14,8 +14,8 @@ import javax.persistence.*;
 @Table(name = "users")
 public class User {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String username;
     private String password;
-    private Integer roleId;
+    private Long roleId;
 }
